refactor(admin): split opencode-providers.ts into focused modules
Replace the 358-line $lib/server/opencode-providers.ts with a focused
module set under $lib/server/opencode/:

- http.ts (29 LOC) — opencodeFetch transport
- config.ts (97 LOC) — getCurrentConfig, patchConfig, normalizeProviderConfig,
  setProviderEnabled, RawConfig type
- catalog.ts (199 LOC) — loadProviderPage and the buildProviderViews assembler
- oauth.ts (46 LOC) — startOauthFlowAtBase, finishOauthFlowAtBase
- results.ts (32 LOC) — actionSuccess, actionFailure builders
- index.ts (12 LOC) — back-compat barrel re-exporting the public API

Each module is well under the 200-line target. All admin route imports
updated to point at the new $lib/server/opencode/index.js barrel; the
public API is unchanged so route handlers and tests only changed paths.

Drive-by cleanup: extracted the inline `^[a-z0-9_-]+$` regex in
providers/custom/+server.ts into a named CUSTOM_PROVIDER_ID_PATTERN
constant.

Verified: bun run admin:check (0 errors), bun run admin:test:unit
(503 pass), bun run test (872 pass).
