+++ packages/admin/src/routes/admin/providers/+server.ts
@@ -1,6 +1,6 @@
 import type { RequestHandler } from './$types';
 import { requireAdmin, jsonResponse, getRequestId } from '$lib/server/helpers.js';
-import { loadProviderPage } from '$lib/server/opencode-providers.js';
+import { loadProviderPage } from '$lib/server/opencode/index.js';
 
 export const GET: RequestHandler = async (event) => {
 	const requestId = getRequestId(event);
