+++ packages/admin/src/routes/admin/providers/oauth/start/+server.ts
@@ -4,7 +4,7 @@ import {
 	startOauthFlowAtBase,
 	actionSuccess,
 	actionFailure,
-} from '$lib/server/opencode-providers.js';
+} from '$lib/server/opencode/index.js';
 import { ensureAuthServer } from '$lib/server/opencode-auth-subprocess.js';
 import { asString, extractInputs } from '../../_helpers.js';
