+++ packages/admin/src/routes/admin/providers/toggle/+server.ts
@@ -6,7 +6,7 @@ import {
 	setProviderEnabled,
 	actionSuccess,
 	actionFailure,
-} from '$lib/server/opencode-providers.js';
+} from '$lib/server/opencode/index.js';
 import { asString } from '../_helpers.js';
 
 /**
